scengen-web/requirements.txt
groq
stagehand-sdk
python-dotenv
Pillow
openai 
playwright
transformers
torch